Fix 501 compile errors across 164 files
Systematic fixes for the macosArm64 target compilation:
- Fix sorted/sortBy calls needing Comparable constraints in starlark_map
- Fix Throwable type parameter restriction in ordered_set
- Remove bogus imports throughout (methods wrongly imported as top-level)
- Fix Result unwrapping (getOrThrow instead of !!)
- Add missing override modifiers on StarlarkValue implementations
- Fix type mismatches (AllocValue, AllocFrozenValue, ValueHolder)
- Fix inline function visibility issues
- Fix destructuring on non-data classes
- Add missing abstract member implementations
- Fix FrozenRef atomic API usage for Kotlin Native
- Port missing methods (normalizeForGoldenTests, pop on SmallSet)
- Fix namespace/package resolution for cross-package extensions

Build now succeeds with 0 compile errors on macosArm64 and metadata targets.

+++ src/commonMain/kotlin/io/github/kotlinmania/starlark_kotlin/collections/symbol/Map.kt
@@ -98,14 +98,14 @@ class SymbolMap<T> private constructor(
 
     /** Look up a value by pre-hashed string key. */
     fun getHashedStr(key: Hashed<String>): T? {
-        val hash = key.hash().promote().toLong()
+        val hash = key.hash().promote()
         val bucket = table[hash] ?: return null
         return bucket.firstOrNull { it.first.asStr() == key.key() }?.second
     }
 
     /** Look up a value by pre-hashed string value, using aligned padded string comparison. */
     fun getHashedStringValue(key: Hashed<String>): T? {
-        val hash = key.hash().promote().toLong()
+        val hash = key.hash().promote()
         val bucket = table[hash] ?: return null
         val keyAligned = Symbol.new(key.key()).asAlignedPaddedStr()
         return bucket.firstOrNull { it.first.asAlignedPaddedStr() == keyAligned }?.second
